refactor: AuthConfiguration 클래스 위치 수정 후 @Import 제거
