+++ src/main/java/nextstep/RoomEscapeApplication.java
@@ -1,20 +1,11 @@
 package nextstep;
 
-import auth.AuthConfig;
-import auth.JwtTokenProvider;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.ApplicationContext;
-import org.springframework.context.annotation.AnnotationConfigApplicationContext;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Import;
 
-@SpringBootApplication(scanBasePackages = {
-        "auth",
-        "nextstep"
-})
+@SpringBootApplication
 public class RoomEscapeApplication {
     public static void main(String[] args) {
-;        SpringApplication.run(RoomEscapeApplication.class, args);
+        SpringApplication.run(RoomEscapeApplication.class, args);
     }
 }
