+++ src/main/java/nextstep/WebMvcConfiguration.java
@@ -1,15 +1,10 @@
 package nextstep;
 
 import auth.AdminInterceptor;
-import auth.AuthConfig;
 import auth.JwtTokenProvider;
 import auth.LoginMemberArgumentResolver;
 import lombok.AllArgsConstructor;
-import lombok.RequiredArgsConstructor;
-import org.springframework.context.ApplicationContext;
-import org.springframework.context.annotation.AnnotationConfigApplicationContext;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Lazy;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
